javascript/packs 以下のディレクトリを修正
packs 配下に JS のコードを全て置いていたため、それらを src/ に移動させた。
