+++ app/javascript/src/components/result.vue
