+++ app/javascript/src/main.vue
