+++ app/javascript/src/store/index.ts
