+++ app/javascript/src/types/responseData.ts
