+++ app/javascript/src/types/result.ts
