+++ app/javascript/src/utils/parseResponseData.ts
