+++ app/javascript/src/view/resultsPage.vue
