+++ config/initializers/assets.rb
@@ -11,6 +11,6 @@ Rails.application.config.assets.version = "1.0"
 Rails.application.config.assets.paths << Rails.root.join("node_modules")
 
 # Precompile additional assets.
-# application.ts, application.css, and all non-JS/CSS in the app/assets
+# main.ts, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
 # Rails.application.config.assets.precompile += %w( admin.js admin.css )
