+++ vuex-shim.d.ts
@@ -1,6 +1,6 @@
 import { ComponentCustomProperties } from 'vue';
 import { Store } from 'vuex';
-import { result } from "./app/javascript/packs/types/result";
+import { result } from "./app/javascript/src/types/result";
 
 declare module '@vue/runtime-core' {
   interface ComponentCustomProperties {
